requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
Django==2.0.6
django-appconf==1.0.4
django-compressor==2.4
django-rest-framework==0.1.0
djangorestframework==3.11.1
idna==2.10
pycparser==2.20
pyOpenSSL==19.1.0
pytz==2020.1
rcssmin==1.0.6
requests==2.24.0
rjsmin==1.1.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
